# 🔧 CRITICAL: Restore MCP Transparency P0 Feature