… .claudedirector/integration-protection/MCP_TRANSPARENCY_P0_COMPLETE.md
… .claudedirector/tests/regression/test_mcp_transparency_p0.py